first go at creating routes based on slugs
For now, the code getting the slugs and data
is inside all-bugs because I know that page works
and that it'll definitely run what's in there.

[slug].js is the equivalent of [id].js in the
example in the NextJS docs. Inside all-bugs,
getCritterSlugs returns an array of objects
that looks like: [ {params:{slug: 'slug'}].

Inside [slug].js, getStaticPaths is supposed
to run getCritterSlugs() and use that to create
the different paths. This is structured to match
the example in the nextjs.org/learn/basics tutorial.

A console.log of critterSlugs showed an array of
strings where each string is a slug: ['agrias-butterfly',
'ant', 'atlas-moth'] etc.

A console.log of the getCritterSlugs return shows the
array of objects noted above.

getStaticPaths is *broken* because it's getting
an "Invalid `paths` value returned from
getStaticPaths in /[slug]".

Currently not sure why that's happening. Womp.