Finish dynamic routes and data fetching for each critter
Change “all-bugs” and “all-fish” to be “bugs” and “fish”
so the URLs are consistent with the individual pages
being “/bugs/[slug]” and “/fish/[slug]”. Navigating
back to “All Bugs” from “/bugs/ant” for example will
take the user to “/bugs/“.

Remove routing functions from all-bugs file and split
them into files for bug-data and fish-data, as well as
two [slug] pages - one for /bugs/[slug] and one
for /fish/[slug].

Previously tried to have one query for all critters but
it’s more clear (to me) to keep them separate the way they
are in the schema.

Have individual critter page link to all bugs or all fish,
and have each critter name on those pages be a link to the
critter’s individual page.