first pass at table - wip
First pass at adding a table to the All Bugs page. Currently
able to grab all data but rendering the list of names as Links
is broken - terminal + console not showing errors in JS, but
it's not rendering anything and I'm not sure why.

I realized that the slugs go to /fish and /bugs and to map over
a list of critters and create the links by `/${critter.typename}/
${critter.slug}` I'd need a plural of the type name. That's fine for
fish but for "bugs" I need to add the "s" - the table in the db is
currently named "bug".

I could avoid dealing with this by going to the db and just changing
the type to "bugs" but that feels like cheating?

This came up by trying to put the table in its own component
instead of having a table directly in the bugs page and another
in the fish page, since I'll also need something similar for the
other views.

Tested earlier as a table directly in the bugs page and the links
to the individual critters by `/bugs/${bug.slug}` worked fine, so
I know the <Link> itself isn't the issue. I think.