Add tailwindCSS and UI, create first table
Create a table using tailwind-ui and replace the initial table
component with that. As part of adding tailwind, update npm
packages and add config files for tailwind and postcss per the
tailwind docs. Updated globals.css for now, but given the use
of tailwind-ui, will probably delete and opt for inline tailwind
only.